+++ src/lib/editor/Document.cpp
@@ -5,6 +5,7 @@
 // https://github.com/albeva/fbide
 //
 #include "Document.hpp"
+#include "DocumentManager.hpp"
 #include "Editor.hpp"
 #include "app/Context.hpp"
 #include "config/ConfigManager.hpp"
@@ -29,7 +30,7 @@ auto defaultEolModeFromConfig(Context& ctx) -> EolMode {
 Document::Document(wxWindow* parent, Context& ctx, const DocumentType type)
 : m_ctx(ctx)
 , m_type(type)
-, m_editor(make_unowned<Editor>(parent, ctx, type))
+, m_editor(make_unowned<Editor>(parent, ctx, &ctx.getDocumentManager().getCodeTransformer(), type))
 , m_encoding(defaultEncodingFromConfig(ctx))
 , m_eolMode(defaultEolModeFromConfig(ctx)) {
     if (m_editor) {