+++ src/lib/format/FormatDialog.cpp
@@ -98,7 +98,7 @@ void FormatDialog::create() {
 
     // Preview editor
     vbox("Previw", { .proportion = 1, .border = 0 }, [&] {
-        m_preview = make_unowned<Editor>(currentParent(), m_ctx, DocumentType::FreeBASIC, true);
+        m_preview = make_unowned<Editor>(currentParent(), m_ctx, nullptr, DocumentType::FreeBASIC, true);
         m_preview->SetReadOnly(true);
         m_preview->SetMinSize(wxSize(-1, 200));
         add(m_preview, { .proportion = 1 });